+++ tests/testthat/test-05-regression.R
@@ -5,6 +5,7 @@ if (identical(Sys.getenv("NOT_CRAN"), "true")) {
     source("regression/dtwb.R", TRUE)
     source("regression/family-distmat.R", TRUE)
     source("regression/family-centroids.R", TRUE)
+    source("regression/centroids.R", TRUE)
     source("regression/custom-dist.R", TRUE)
     source("regression/cvis.R", TRUE)
     source("regression/methods.R", TRUE)